Measure every stage with perf_counter_ns; publish p50/p95 and the void disclosure
The previous run reported 0.0 ms for eleven stages, every value a multiple of
15.625 ms -- the Windows time.monotonic tick. Those readings meant the clock
could not resolve the stage, not that the stage was free.

- StageRecorder now uses time.perf_counter_ns and accumulates in integer
  nanoseconds, converting to milliseconds only at the boundary. RunTiming
  records the clock name, its advertised resolution and the host platform, so a
  latency table can no longer be read without knowing what measured it.
- Measured durations stop being truncated to whole integer milliseconds in
  ToolTrace, ProviderCallTrace and audit events; a 0.4 ms call recorded as 0.
- Every stage now reports p50 and p95, not a lone median, plus min/mean/stdev/max.
- Wall clock, provider-call time and local compute are reported as three
  separate series and are never summed.
- Stage spans are not a partition of a run, so the unattributed remainder and
  the slowest run's stage breakdown are published rather than left implicit.
- New resolveflow.eval.verify_checksums re-hashes every row of a published
  SHA256SUMS manifest against disk, and flags artifacts on disk that no row
  covers.
- The methodology and the results page now carry the voided-run disclosure and
  state plainly that no live Cohere run has been performed since the token
  budget fix, so no model-dependent number in this repository is measured.
- FixtureChatAdapter is documented as what it is: a hand-written stub that
  cannot be prompt-injected, not a recording of model output.
- Adds docs/THREAT_MODEL.md (attacker position, capability, goals, oracles,
  and what is not tested) and docs/COHERE-POSITIONING.md.

Fixture A/B re-run on Linux under the fixed budget: 32/32 runs complete, zero
token_budget_exhausted, zero stages reading 0.0 ms, 39/39 checksums verified.

+++ python/resolveflow/agent/contracts.py
@@ -103,7 +103,7 @@ class ProviderTrace(FrozenModel):
     tool_call_names: tuple[str, ...]
     citation_ids: tuple[str, ...]
     usage: ProviderUsage
-    duration_ms: int = Field(ge=0)
+    duration_ms: float = Field(ge=0.0)
     safe_error_code: str | None = None
 
 
@@ -113,7 +113,7 @@ class ToolTrace(FrozenModel):
     status: Literal["ok", "rejected", "timeout", "error"]
     authorization: Literal["allowed", "denied", "not_evaluated"]
     arguments_hash: str
-    duration_ms: int = Field(ge=0)
+    duration_ms: float = Field(ge=0.0)
     provenance_ids: tuple[str, ...]
     safe_error_code: str | None = None
     external_write: Literal[False] = False
